Moved object state check methods.
