+++ src/engine/core/managers/interface/notifications/NotificationManager.ts
@@ -28,9 +28,9 @@ import {
 import { GlobalSoundManager } from "@/engine/core/managers/sounds/GlobalSoundManager";
 import { Stalker } from "@/engine/core/objects";
 import { abort, assert } from "@/engine/core/utils/assertion";
-import { isHeavilyWounded } from "@/engine/core/utils/check/check";
 import { isStalkerClassId } from "@/engine/core/utils/check/is";
 import { LuaLogger } from "@/engine/core/utils/logging";
+import { isHeavilyWounded } from "@/engine/core/utils/object";
 import { getInventoryNameForItemSection } from "@/engine/core/utils/spawn";
 import { captions, TCaption } from "@/engine/lib/constants/captions/captions";
 import { scriptSounds } from "@/engine/lib/constants/sound/script_sounds";
