+++ src/engine/core/objects/state/state/EvaluatorStateIdleAlife.ts
@@ -2,9 +2,9 @@ import { action_planner, LuabindClass, property_evaluator } from "xray16";
 
 import { StalkerStateManager } from "@/engine/core/objects/state/StalkerStateManager";
 import { EStalkerState, EStateEvaluatorId } from "@/engine/core/objects/state/types";
-import { EActionId } from "@/engine/core/schemes";
-import { isObjectMeeting } from "@/engine/core/utils/check/check";
+import { EActionId } from "@/engine/core/schemes/base";
 import { LuaLogger } from "@/engine/core/utils/logging";
+import { isObjectMeeting } from "@/engine/core/utils/object";
 import { Optional } from "@/engine/lib/types";
 
 const logger: LuaLogger = new LuaLogger($filename);
