+++ src/engine/core/schemes/combat_ignore/CombatProcessEnemyManager.ts
@@ -4,8 +4,8 @@ import { registry } from "@/engine/core/database";
 import { SmartTerrain } from "@/engine/core/objects/server/smart_terrain/SmartTerrain";
 import { AbstractSchemeManager } from "@/engine/core/schemes";
 import { ISchemeCombatIgnoreState } from "@/engine/core/schemes/combat_ignore/index";
-import { isObjectEnemy } from "@/engine/core/utils/check/check";
 import { LuaLogger } from "@/engine/core/utils/logging";
+import { isObjectEnemy } from "@/engine/core/utils/object";
 import { logicsConfig } from "@/engine/lib/configs/LogicsConfig";
 import { ACTOR_ID } from "@/engine/lib/constants/ids";
 import { MAX_U16 } from "@/engine/lib/constants/memory";
