+++ src/engine/core/schemes/corpse_detection/evaluators/EvaluatorCorpseDetect.ts
@@ -3,8 +3,8 @@ import { level, LuabindClass, property_evaluator } from "xray16";
 import { IRegistryObjectState, registry } from "@/engine/core/database";
 import { IReleaseDescriptor, ReleaseBodyManager } from "@/engine/core/managers/world/ReleaseBodyManager";
 import { ISchemeCorpseDetectionState } from "@/engine/core/schemes/corpse_detection";
-import { isObjectWounded } from "@/engine/core/utils/check/check";
 import { isLootableItem } from "@/engine/core/utils/check/is";
+import { isObjectWounded } from "@/engine/core/utils/object";
 import { communities } from "@/engine/lib/constants/communities";
 import { ClientObject, Optional, TDistance, TNumberId, Vector } from "@/engine/lib/types";
