+++ src/engine/core/schemes/help_wounded/evaluators/EvaluatorWoundedExist.ts
@@ -3,8 +3,8 @@ import { level, LuabindClass, property_evaluator } from "xray16";
 import { IRegistryObjectState, registry } from "@/engine/core/database";
 import { ISchemeHelpWoundedState } from "@/engine/core/schemes/help_wounded";
 import { ISchemeWoundedState } from "@/engine/core/schemes/wounded";
-import { isObjectWounded } from "@/engine/core/utils/check/check";
 import { LuaLogger } from "@/engine/core/utils/logging";
+import { isObjectWounded } from "@/engine/core/utils/object";
 import { communities } from "@/engine/lib/constants/communities";
 import { ClientObject, EScheme, TDistance, TNumberId, Vector } from "@/engine/lib/types";
