+++ src/engine/core/schemes/ph_minigun/MinigunManager.ts
@@ -4,11 +4,15 @@ import { getObjectByStoryId, registry } from "@/engine/core/database";
 import { AbstractSchemeManager } from "@/engine/core/schemes/base";
 import { ISchemeMinigunState } from "@/engine/core/schemes/ph_minigun/ISchemeMinigunState";
 import { abort } from "@/engine/core/utils/assertion";
-import { isHeavilyWounded } from "@/engine/core/utils/check/check";
 import { isActiveSection } from "@/engine/core/utils/check/is";
 import { pickSectionFromCondList, TConditionList } from "@/engine/core/utils/ini";
-import { isMonsterScriptCaptured, scriptReleaseMonster } from "@/engine/core/utils/scheme";
-import { switchObjectSchemeToSection, trySwitchToAnotherSection } from "@/engine/core/utils/scheme/switch";
+import { isHeavilyWounded } from "@/engine/core/utils/object";
+import {
+  isMonsterScriptCaptured,
+  scriptReleaseMonster,
+  switchObjectSchemeToSection,
+  trySwitchToAnotherSection,
+} from "@/engine/core/utils/scheme";
 import { createEmptyVector, createVector, yaw } from "@/engine/core/utils/vector";
 import { ACTOR, NIL } from "@/engine/lib/constants/words";
 import { Car, ClientObject, Optional, TName, TSection, TStringId, TTimestamp, Vector } from "@/engine/lib/types";
