+++ src/engine/core/utils/object/index.ts
@@ -1,3 +1,4 @@
+export * from "@/engine/core/utils/object/object_danger";
 export * from "@/engine/core/utils/object/object_general";
 export * from "@/engine/core/utils/object/object_location";
 export * from "@/engine/core/utils/object/object_state";
